Refactor Node from Tree into its own component
